+++ database/migrations/2024_05_30_121543_create_article_destinations_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('slug');
             $table->string('image')->nullable();
             $table->text('body');
-            $table->integer('ratting_article_acc');
+            $table->double('ratting_article_acc');
 
             $table->softDeletes();
             $table->timestamps();
